Fix up to get both sides to confirm goals, fouls and faceoffs
It appears that it could be possible for one side of a match to think
a foul occurred and exactly the same time as the other side thought a
goal had occurred.  This means each side gets out of step - and then
eventually you get to a point where both sides thinks the other should
serve.  At this point the whole game hangs

To get round this, we try and confirm when two things happen, with the
master end being the arbiter.