Fix problem where messages one after the other were overwriting the option buffer.
I think this problem has been the route cause of message failures -
since it attempts to re-use an already running writes options.  If
that previous write was already in a queue (chained) then it will be
over written - one write will be lost, and other will be repeated
twice.  This could have been the cause of additional goals.

Lets see