Adjust placement of turning foreign key pragma off and on
It seems that placing a pragma foriegn key off, after preparing a
statement causes it to fail when executing (this is only a possibility
from reading the sqlite documentation).  Either that, or the fact its
a no-op inside a transaction was causing the INSERT OR REPLACE to
insert a new user to fail.  Moving the pragmas outside of the
transaction boundary seems to sort the problem