Thinking that sending $ down pipe to unblock it is a bad thing. So stop it
I have been sending a '$' down the pipe as an unblock signal.  But I
an wondering if that causes the next read to be satisfied with a zero
length string.  I am taking it away in this commit so as to prove
whether that is the case or not.

Also setting all the modes to 'b'.  This should be the default, but it
might be safer to specify it explicitly.