Fixes comms timeout at start of match
The stream read function was returning the time of the message as a
text string rather than as an integer.  This caused delay calculations
to be wrong, and the whole thing just timed out.  By making it return
an int, the calculations can now be performed correctly.

Also removed the parameter from the match.start function.  I think
this is legacy code that should have been removed before