Fix issue where a foul by me screwed computer trying to hit puck
This fixes a problem where by a foul by me does not reset the puck
place and move to hit time delays, so a possible previous version of
the delay is used - causing the computer to foul.  This is sad,
because it gives the player an unfair advantage - doing a foul causes
the computer to foul - giving you control of the puck

We now reset the move to hit and puck delay timers if there is a new
request to serve.